Temp hack to fix grep results on DOS EOL files.
When viewing grep results for files containing DOS EOL (i.e. CRLF)
on Mac, there is a trailing ^M for every matching line. Wgrep checks
that the file contents at the matching line are the same as the
original value from the grep results. This will never pass because
of contradictory "coding systems" in the grep output and the
opened file. The opened buffer will detect the DOS EOL and hide
the trailing ^M control character.